Major refactoring for skinning continues, helmet and blend-shape scenes export again, and skinning scene doesn't give validation errors, but is incomplete

+++ src/MeshVertices.h
@@ -68,7 +68,7 @@ inline VertexComponents floats(gsl::span<T> span)
 template<typename T>
 inline VertexComponents shorts(gsl::span<T> span)
 {
-	return VertexComponents(reinterpret_span<float>(span));
+	return VertexComponents(reinterpret_span<ushort>(span));
 }
 
 typedef std::vector<VertexComponents> VertexElementsPerSetIndex;
